use tap-tester logger and base case inheritance
