+++ tests/test_config.py
@@ -1,4 +1,6 @@
 import os
+from tap_tester.logger import LOGGER
+
 
 potential_paths = [
     'tests/',
@@ -20,33 +22,33 @@ def go_to_tests_directory():
 ##########################################################################
 
 
-print("Acquiring path to tests directory.")
+LOGGER.info("Acquiring path to tests directory.")
 cwd = go_to_tests_directory()
 
-print("Reading in filenames from tests directory.")
+LOGGER.info("Reading in filenames from tests directory.")
 files_in_dir = os.listdir(cwd)
 
-print("Dropping files that are not of the form 'test_<feature>.py'.")
+LOGGER.info("Dropping files that are not of the form 'test_<feature>.py'.")
 test_files_in_dir = [fn for fn in files_in_dir if fn.startswith('test_') and fn.endswith('.py')]
 
-print("Dropping test_client.py from test files.")
+LOGGER.info("Dropping test_client.py from test files.")
 if 'test_client.py' in test_files_in_dir:
     test_files_in_dir.remove('test_client.py')
 
-print("Files found: {}".format(test_files_in_dir))
+LOGGER.info("Files found: {}".format(test_files_in_dir))
 
-print("Reading contents of circle config.")
+LOGGER.info("Reading contents of circle config.")
 with open(cwd + "/../.circleci/config.yml", "r") as config:
     contents = config.read()
 
-print("Parsing circle config for run blocks.")
+LOGGER.info("Parsing circle config for run blocks.")
 runs = contents.replace(' ', '').replace('\n', '').split('-run:')
 
-print("Verify all test files are executed in circle...")
+LOGGER.info("Verify all test files are executed in circle...")
 tests_not_found = set(test_files_in_dir)
 for filename in test_files_in_dir:
-    print("\tVerifying {} is running in circle.".format(filename))
+    LOGGER.info("\tVerifying {} is running in circle.".format(filename))
     if any([filename in run for run in runs]):
         tests_not_found.remove(filename)
 assert tests_not_found == set(), "The following tests are not running in circle:\t{}".format(tests_not_found)
-print("\t SUCCESS: All tests are running in circle.")
+LOGGER.info("\t SUCCESS: All tests are running in circle.")
